feat: Add Firebase configuration and image upload functionality
This commit adds the necessary Firebase configuration and image upload functionality to the AddProductModal component. The firebaseConfig.js file is created to initialize the Firebase app and storage. The handleImageChange function is implemented to handle the selection of an image file, and the handleImageUpload function is added to handle the upload of the selected image to Firebase storage. The uploaded image URL is then used to update the imageUrl state. Additionally, the necessary imports and state variables are added to support the image upload functionality.

Refactor the VendorNav and VendorDashboard components
Merge branch 'dev' of https://github.com/Janani-Mayadunna/EAD_web_online_shop into vendor-mng
Refactor the VendorCategory.css file to update badge colors
Update error message styling in AdminLogin and CSRLogin components
Sort categories and products by _id in descending order